Worked on Character entity and file system
!! CODE PROBABLY WON'T WORK !!
I used Visual Studio Code so I haven't tested this recent commit, so assume it doesn't :)

Currently I have it so all entity related information is stored in one central file, but I think it might be worth to create system that allows for each entity to have their own individual files. I'll need to create some method in Entity for this.