Started on rotational collision
Thoughts: It's going super fast because the collision method doesn't take the change in velocity into account. since angAccel is very large, there is a huge difference in velocity between times. Not accounting for this is messing things up. When it is fixed this should be applied to the linear collision checker as well.